Module aws_sdk_cognitosync::error

Modules

already_streamed_exception

See AlreadyStreamedException

concurrent_modification_exception

See ConcurrentModificationException

duplicate_request_exception

See DuplicateRequestException

internal_error_exception

See InternalErrorException

invalid_configuration_exception

See InvalidConfigurationException

invalid_lambda_function_output_exception

See InvalidLambdaFunctionOutputException

invalid_parameter_exception

See InvalidParameterException

lambda_throttled_exception

See LambdaThrottledException

limit_exceeded_exception

See LimitExceededException

not_authorized_exception

See NotAuthorizedException

resource_conflict_exception

See ResourceConflictException

resource_not_found_exception

See ResourceNotFoundException

too_many_requests_exception

See TooManyRequestsException

Structs

AlreadyStreamedException

An exception thrown when a bulk publish operation is requested less than 24 hours after a previous bulk publish operation completed successfully.

BulkPublishError

ConcurrentModificationException

Thrown if there are parallel requests to modify a resource.

DeleteDatasetError

DescribeDatasetError

DescribeIdentityPoolUsageError

DescribeIdentityUsageError

DuplicateRequestException

An exception thrown when there is an IN_PROGRESS bulk publish operation for the given identity pool.

GetBulkPublishDetailsError

GetCognitoEventsError

GetIdentityPoolConfigurationError

InternalErrorException

Indicates an internal service error.

InvalidConfigurationException

InvalidLambdaFunctionOutputException

The AWS Lambda function returned invalid output or an exception.

InvalidParameterException

Thrown when a request parameter does not comply with the associated constraints.

LambdaThrottledException

AWS Lambda throttled your account, please contact AWS Support

LimitExceededException

Thrown when the limit on the number of objects or operations has been exceeded.

ListDatasetsError

ListIdentityPoolUsageError

ListRecordsError

NotAuthorizedException

Thrown when a user is not authorized to access the requested resource.

RegisterDeviceError

ResourceConflictException

Thrown if an update can’t be applied because the resource was changed by another call and this would result in a conflict.

ResourceNotFoundException

Thrown if the resource doesn’t exist.

SetCognitoEventsError

SetIdentityPoolConfigurationError

SubscribeToDatasetError

TooManyRequestsException

Thrown if the request is throttled.

UnsubscribeFromDatasetError

UpdateRecordsError

Enums

BulkPublishErrorKind

DeleteDatasetErrorKind

DescribeDatasetErrorKind

DescribeIdentityPoolUsageErrorKind

DescribeIdentityUsageErrorKind

GetBulkPublishDetailsErrorKind

GetCognitoEventsErrorKind

GetIdentityPoolConfigurationErrorKind

ListDatasetsErrorKind

ListIdentityPoolUsageErrorKind

ListRecordsErrorKind

RegisterDeviceErrorKind

SetCognitoEventsErrorKind

SetIdentityPoolConfigurationErrorKind

SubscribeToDatasetErrorKind

UnsubscribeFromDatasetErrorKind

UpdateRecordsErrorKind