Module aws_sdk_cognitosync::input

Modules

bulk_publish_input

See BulkPublishInput

delete_dataset_input

See DeleteDatasetInput

describe_dataset_input

See DescribeDatasetInput

describe_identity_pool_usage_input

See DescribeIdentityPoolUsageInput

describe_identity_usage_input

See DescribeIdentityUsageInput

get_bulk_publish_details_input

See GetBulkPublishDetailsInput

get_cognito_events_input

See GetCognitoEventsInput

get_identity_pool_configuration_input

See GetIdentityPoolConfigurationInput

list_datasets_input

See ListDatasetsInput

list_identity_pool_usage_input

See ListIdentityPoolUsageInput

list_records_input

See ListRecordsInput

register_device_input

See RegisterDeviceInput

set_cognito_events_input

See SetCognitoEventsInput

set_identity_pool_configuration_input

See SetIdentityPoolConfigurationInput

subscribe_to_dataset_input

See SubscribeToDatasetInput

unsubscribe_from_dataset_input

See UnsubscribeFromDatasetInput

update_records_input

See UpdateRecordsInput

Structs

BulkPublishInput

The input for the BulkPublish operation.

DeleteDatasetInput

A request to delete the specific dataset.

DescribeDatasetInput

A request for meta data about a dataset (creation date, number of records, size) by owner and dataset name.

DescribeIdentityPoolUsageInput

A request for usage information about the identity pool.

DescribeIdentityUsageInput

A request for information about the usage of an identity pool.

GetBulkPublishDetailsInput

The input for the GetBulkPublishDetails operation.

GetCognitoEventsInput

A request for a list of the configured Cognito Events

GetIdentityPoolConfigurationInput

The input for the GetIdentityPoolConfiguration operation.

ListDatasetsInput

Request for a list of datasets for an identity.

ListIdentityPoolUsageInput

A request for usage information on an identity pool.

ListRecordsInput

A request for a list of records.

RegisterDeviceInput

A request to RegisterDevice.

SetCognitoEventsInput

A request to configure Cognito Events"
"

SetIdentityPoolConfigurationInput

The input for the SetIdentityPoolConfiguration operation.

SubscribeToDatasetInput

A request to SubscribeToDatasetRequest.

UnsubscribeFromDatasetInput

A request to UnsubscribeFromDataset.

UpdateRecordsInput

A request to post updates to records or add and delete records for a dataset and user.