Module aws_sdk_cognitosync::model

Modules

cognito_streams

See CognitoStreams

dataset

See Dataset

identity_pool_usage

See IdentityPoolUsage

identity_usage

See IdentityUsage

push_sync

See PushSync

record

See Record

record_patch

See RecordPatch

Structs

CognitoStreams

Configuration options for configure Cognito streams.

Dataset

A collection of data for an identity pool. An identity pool can have multiple datasets. A dataset is per identity and can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don’t exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

IdentityPoolUsage

Usage information for the identity pool.

IdentityUsage

Usage information for the identity.

PushSync

Configuration options to be applied to the identity pool.

Record

The basic data structure of a dataset.

RecordPatch

An update operation for a record.

Enums

BulkPublishStatus

Operation

Platform

StreamingStatus