Module aws_sdk_cognitosync::operation

Structs

BulkPublish

Initiates a bulk publish of all existing datasets for an Identity Pool to the configured stream. Customers are limited to one successful bulk publish per 24 hours. Bulk publish is an asynchronous request, customers can see the status of the request via the GetBulkPublishDetails operation.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

DeleteDataset

Deletes the specific dataset. The dataset will be deleted permanently, and the action can't

DescribeDataset

Gets meta data about a dataset by identity and dataset name. With Amazon Cognito Sync, each

DescribeIdentityPoolUsage

Gets usage details (for example, data storage) about a particular identity pool.

DescribeIdentityUsage

Gets usage information for an identity, including number of datasets and data usage.

GetBulkPublishDetails

Get the status of the last BulkPublish operation for an identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

GetCognitoEvents

Gets the events and the corresponding Lambda functions associated with an identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

GetIdentityPoolConfiguration

Gets the configuration settings of an identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

ListDatasets

Lists datasets for an identity. With Amazon Cognito Sync, each identity has access only to

ListIdentityPoolUsage

Gets a list of identity pools registered with Cognito.

ListRecords

Gets paginated records, optionally changed after a particular sync count for a dataset and

RegisterDevice

Registers a device to receive push sync notifications.

This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.

SetCognitoEvents

Sets the AWS Lambda function for a given event type for an identity pool. This request only updates the key/value pair specified. Other key/values pairs are not updated. To remove a key value pair, pass a empty value for the particular key.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

SetIdentityPoolConfiguration

Sets the necessary configuration for push sync.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

SubscribeToDataset

Subscribes to receive notifications when a dataset is modified by another device.

This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.

UnsubscribeFromDataset

Unsubscribes from receiving notifications when a dataset is modified by another device.

This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.

UpdateRecords

Posts updates to records and adds and deletes records for a dataset and user.